IN PROGRESS - issue SPAGOBI-1616: As SpagoBI User I want that when the dataset cache is full the dataset manager free some space deleting old objects from it https://www.spagoworld.org/jira/browse/SPAGOBI-1616
git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi/V_4.x/Server/trunk@22008 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ SpagoBICockpitEngine/src/it/eng/spagobi/dataset/cache/ICacheMetadata.java
@@ -42,7 +42,7 @@ public interface ICacheMetadata {
 	/**
 	 * @return the number of the objects cached
 	 */
-	Double getNumberOfObjects();
+	Integer getNumberOfObjects();
 	
 	/**
 	 * @return true if the cache space is near the full limit
